requirements.txt
anthropic>=0.40.0
flask==3.0.3
flask-sqlalchemy==3.1.1
apscheduler==3.10.4
python-dotenv==1.0.1
requests==2.31.0
sqlalchemy==2.0.30
